refactor dashboard to have the home tab
