+++ src/sql/parts/dashboard/common/dashboardPage.component.html
@@ -10,9 +10,11 @@
 			<dashboard-widget-wrapper #properties *ngIf="propertiesWidget" [_config]="propertiesWidget" style="padding-left: 10px; padding-right: 10px; height: 90px; display: block">
 			</dashboard-widget-wrapper>
 		</div>
-		<div [ngGrid]="gridConfig">
-			<dashboard-widget-wrapper *ngFor="let widget of widgets" [(ngGridItem)]="widget.gridItemConfig" [_config]="widget">
-			</dashboard-widget-wrapper>
-		</div>
+		<panel class="fullsize">
+			<tab *ngFor="let tab of tabs" [title]="tab.name" class="fullsize" [identifier]="tab.id">
+				<dashboard-tab [tab]="tab">
+				</dashboard-tab>
+			</tab>
+		</panel>
 	</div>
 </div>
