+++ src/sql/parts/dashboard/common/dashboardTab.component.html
@@ -1,0 +1,10 @@
+<!--
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the Source EULA. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+-->
+<div [ngGrid]="gridConfig">
+	<dashboard-widget-wrapper *ngFor="let widget of tab.widgets" [(ngGridItem)]="widget.gridItemConfig" [_config]="widget">
+	</dashboard-widget-wrapper>
+</div>
